Improve registering unknown atoms
For one, don't override mrcdict with a generic dict.
Secondly, document how mrc and python handle missing items
inside the dictionary.

Introduces the base_class for mrcdict.
This enables other atoms to define their own base_class as a
default for an unkown item. Which is used in the drefAtom.

Introduces the GobbleAtom, which gobbles all data inside an atom.
This makes it possible to use this as a default atom as to continue
parsing.