Add panorama atom support to mr_quicktime
For QTVR panoramas there are two special atoms
One is the PanoMediaInfo `pInf` atom.
    Contains the name of the scene, defaults and
    a table to map nodeIDs to time.

The other is the `pano` as the sample description.
    Contains information that is shared by many nodes.
    Note: reserved2 and majorVersion should be 0,
        sometimes have a value of 1.

Improves handling of two other atoms: gmin and gmhd, which weren't
fleshed out yet.

VideoSampleDescription was also extended by having a
sampleDescriptionTable which will now choose the right SampleDescription
based on the type/data_format.