PANORAMA: add hotspot decoding
Hotspots are encoded in the SMC format with a depth of 8 bits.
SMC uses a palette for the encoding. Due to use only leveraging a
tiny part of FFmpeg, FFmpeg doesn't know which palette to use.

In theory it's possible for hotspots to use a different encoding
and a different palette. In practise we haven't seen it so far.

The palette in qt_palette is taken from FFmpeg.