Adds a rudimentary 3D viewer for panorama qtvr1
Rudimentary, it works, but it's not perfect.
It projects the image on a cylinder, that is distinctly different
than how QTVR rendered them in 1995.
Read the doc message in the viewer3d file for more info on how to run.

A quick overview: the images from make_mosaic are projected on a
cylinder and a custom camera object is used to reflect how the original
works. pygfx and wgpu were chosen for their ease of use with cylinders.