VIEWER: improve camera rotation
Fixes bug where movement was almost impossible at an Y angle of 90
degrees.

vecx and vecy are a tuple of length three that denote the position
where the mouse was on mouse down. Irrelevant for the movement that
comes after. The movement relative to where the mouse was clicked
is all in delta. delta[0] for movement in the horizontal pane and
delta[1] for movement in the vertical pane.